Consolidate OTel serialization helpers and operation-error recording across clients
Merge the per-client `OtelContext` source-gen partials into a single
`Common/OtelContext.cs` that registers the union of all message-part types
serialized by both the chat and realtime clients. Chat-only and realtime-only
serialization POCOs that the merged context must reference are hoisted from
nested private types to top-level `internal sealed` classes (still co-located
with their owning client file via an SA1402 pragma).

Move the chat-side serialization helpers — `SerializeChatMessages`,
`_defaultOptions`, `_emptyObject`, `CreateDefaultOptions`,
`DeriveModalityFromMediaType`, and `ExtractCodeFromInputs` — into
`Common/OtelMessageSerializer.cs`. `OpenTelemetryImageGenerator` and
`OpenTelemetrySpeechToTextClient` now call the shared serializer directly
rather than reaching into `OpenTelemetryChatClient` for it.
`OpenTelemetryRealtimeClientSession` also drops its near-duplicate
`DeriveModalityFromMediaType` and routes its three blob/uri/file emission
sites through `OtelMessageSerializer.DeriveModalityFromMediaType`. The two
implementations were already functionally equivalent across all inputs
(null, empty, no slash, malformed, with parameters); the kept span-based
form runs in a single pass instead of three sequential `StartsWith` checks.

Add `OpenTelemetryLog.RecordOperationError(Activity?, ILogger?, Exception?)`
in `Common/` to consolidate the eight-line "set error tag, set error status,
log via ILogger" block that was previously duplicated across all seven
OpenTelemetry* clients. `OpenTelemetryHostedFileClient.SetErrorStatus` reduces
to a one-line expression-bodied delegator.

No behavioral change. Build is clean across all TFMs and all 669 tests pass
per TFM.
