+++ src/Libraries/Microsoft.Extensions.AI/Common/OpenTelemetryLog.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Diagnostics;
 using Microsoft.Extensions.Logging;
 
 namespace Microsoft.Extensions.AI;
@@ -14,4 +15,24 @@ internal static partial class OpenTelemetryLog
         Level = LogLevel.Warning,
         Message = "gen_ai.client.operation.exception")]
     internal static partial void OperationException(ILogger logger, Exception error);
+
+    /// <summary>Stamps the operation error tag/status on <paramref name="activity"/> and logs the exception.</summary>
+    /// <remarks>No-op when <paramref name="error"/> is <see langword="null"/>.</remarks>
+    internal static void RecordOperationError(Activity? activity, ILogger? logger, Exception? error)
+    {
+        if (error is null)
+        {
+            return;
+        }
+
+        _ = activity?
+            .AddTag(OpenTelemetryConsts.Error.Type, error.GetType().FullName)
+            .SetStatus(ActivityStatusCode.Error, error.Message);
+
+        if (logger is not null)
+        {
+            OperationException(logger, error);
+        }
+    }
 }
+
